% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_function.R
\docType{methods}
\name{autoPCAplotter}
\alias{autoPCAplotter}
\title{autoPCAplotter}
\usage{
autoPCAplotter(
  se,
  pca = NA,
  var_explained = NA,
  MAX_CLASSES_FOR_PCA = 8,
  exclude = c(),
  col_set_factor = "Set1",
  col_set_numeric = "Blues",
  plot_numeric = F
)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{pca}{pca results as outputted by prcomp, by default this is NA and the function calculates the principle components based on assays(se)$final}

\item{var_explained}{variance explained by each principle component, by default this is NA and calculated based on the final data}

\item{MAX_CLASSES_FOR_PCA}{Maximum number of classes to consider}

\item{exclude}{An array of covariate names that should be excluded}

\item{col_set_factor}{RColorBrewer color scheme for factors}

\item{col_set_numeric}{RColorBrewer color scheme for numbers}

\item{plot_numeric}{Flag to indicate whether numberic covariates should be plotted as well}
}
\description{
Autoplotter that plots a PCA for each selected covariate, which are those that have more than 1 class and less
than MAX_CLASSES_FOR_PCA. Will also plot numeric data.
}
