% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_plots.R
\name{BOR_boxplot}
\alias{BOR_boxplot}
\title{BOR_boxplot}
\usage{
BOR_boxplot(
  marker,
  bor,
  label,
  ylab = NA,
  title = paste(label, "by BOR"),
  jitter_method = "jitter",
  calculate_p_value = T,
  COLOR_SCHEME = BOR_colors(),
  overlay_dot_col = "#444444",
  legend_pos = "topright",
  legend_cex = 0.7,
  cex = 1,
  ylim = NULL,
  resp_classes = c("CR", "PR"),
  non_resp_classes = c("PD", "UNK")
)
}
\arguments{
\item{marker}{vector with a numeric marker value, eg. gene expression or signature expression}

\item{bor}{vector with BOR values (CR, PR, SD. NCRNPD, PD, UNK)}

\item{label}{Label used in the title}

\item{ylab}{(Default: value that is used for label) y axis label}

\item{title}{Title of the plot can be specified or is automatically derived from the y label.}

\item{jitter_method}{(Default: jitter) Stripchart overlay of the boxplots. Can also be overplot and stack}

\item{calculate_p_value}{(Default: TRUE) Should a Wilcoxon p-value between CR/PR vs. PD be calculated?}

\item{COLOR_SCHEME}{(Default: BOR_colors()) Names BOR color array}

\item{overlay_dot_col}{(Default: "#444444") Stripchart color}

\item{legend_pos}{(Default: 'topright') Position of the legend}

\item{legend_cex}{(Defaul: 0.7) cex of the legend text}

\item{cex}{font_sizes}

\item{ylim}{ylim}

\item{resp_classes}{(Default: c('CR','PR')) When calculating a p-value two groups need to be compared. Depending on the context only CR, CR/PR or even CR/PR/SD are used as responders.}

\item{non_resp_classes}{(Default: c('PD','UNK')) When calculating a p-value two groups need to be compared. Depending on the context different groups are used as non-responders}
}
\description{
Standard boxplot that splits a marker by BOR levels, makes boxplots and calculates a p-value
}
