% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_heatmap.R
\name{complex_heatmap}
\alias{complex_heatmap}
\title{complex_heatmap}
\usage{
complex_heatmap(
  se,
  selector = NULL,
  genes,
  signatures = NULL,
  ihc = NULL,
  cluster_rows = TRUE,
  cluster_cols = NULL,
  covs = NULL,
  annotation_colors = NA,
  column_order = NULL,
  barplot_color = NA,
  barplot_legend = "grey",
  pathway_projection = "means",
  rowscale = TRUE,
  rowscale_cap = 3,
  barplot_decreasing_order = TRUE,
  barplot_specific_order = NULL,
  barplot_censor = NA,
  barplot_height = 30,
  labels_col = NULL,
  gaps_col = NULL,
  custom_legend = NULL,
  fontsize = 8,
  pathways_name_cap = 40,
  main
)
}
\arguments{
\item{se}{SummarizedExperiment cube that contains all the data}

\item{selector}{logic vector with length of samples that should be included. This is mainly used in conjunction with GSVA as gene set projection method, because it works better with a larger set of samples}

\item{genes}{Genes to be displayed, which need to be contained in the SE cube}

\item{signatures}{List of gene vectors indicating different signatures / pathways / gene sets}

\item{ihc}{Array of names in the colData data.frame of the SE that include IHC}

\item{cluster_rows}{logic flag indicating whether the rows of the three data panels should be clustered. This is done using hierarchical clustering by Ward's method and using the optimal leaf ordering algorithm as implemented in the cba package.}

\item{cluster_cols}{logic flag indicating whether the samples should be clustered. Is ignored if a barplot is specificed and is automatically set to TRUE if not specified. If you want to impose a custom order order the matrix beforehand and set it to FALSE.}

\item{covs}{Array of names in the colData data.frame of the SE indicating additional covariates that should be displayed on top of the three panels}

\item{annotation_colors}{(default: NA) List of colors for the covariates that are shown on top.}

\item{column_order}{Indicating the covariate (name of the colData column) that the samples should be ordered by. This numberic value is used for the top barplot.}

\item{barplot_color}{This can be either NA - which colors all bars the color specified in barplot_legend, or a factorial where the levels need to correspond to the names of the barplot legend.}

\item{barplot_legend}{This can be a single color indicating the colors of all bars or a color scheme, which is an array of colors with names that correspond to the levels indicated in barplot_color}

\item{pathway_projection}{Pathway projection method. 'means' just averages the gene expression values of the passed gene sets, 'gsva' uses the R/Bioconductor package GSVA (using a Gaussian kernel in log2 space)}

\item{rowscale}{Logic flag indicating whether each row should be row scaled}

\item{rowscale_cap}{Caps the values in all three panels to +/- the cap value (default: 3), which avoids that the color schemes are driven by outliers. A values of 3 indicates that only values within 3 standard deviations are shown.}

\item{barplot_decreasing_order}{Logic flag that indicates the sorting order of the barplot. Can be NA - no sorting}

\item{barplot_specific_order}{Instead of sorting a specific ordering can be passed in case there are multiple indications or datasets that should be grouped (This overrides the barplot_decreasing_order flag)}

\item{barplot_censor}{Censor vector if the barplot is used for survival data}

\item{barplot_height}{Height of the barplot in grub small px (default: 30)}

\item{labels_col}{Custom sample labels. By default these are the column labels of the passed SE}

\item{gaps_col}{array of indices that indicates where a gap in the columns should be placed}

\item{custom_legend}{Changes the labels for the color scales}

\item{fontsize}{Used fontsize in the heatmap}

\item{pathways_name_cap}{Pathway labels are be capped at 40}

\item{main}{Title for the complex heatmap}
}
\description{
Complex heatmap function that shows a barplot on top, can show several covariates and
can display 3 panels of data - IHC, gene expression and gene sets
}
