% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liver_confounding.R
\docType{methods}
\name{deriveLiverScore}
\alias{deriveLiverScore}
\title{deriveLiverScore}
\usage{
deriveLiverScore(dat, method = c("mean", "gsva"))
}
\arguments{
\item{dat}{Gene expression matrix - rows are genes and columns are samples}

\item{method}{Method used to estimate the liver score ('mean' or 'gsva')}
}
\value{
Vector of liver scores indicating the levels of liver confounding
}
\description{
Derive a liver score for each sample based on the 22 liver gene signature
}
