% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_gff.R
\name{extract_attributes}
\alias{extract_attributes}
\title{Extract gtf attribute}
\usage{
extract_attributes(gtf_attributes, att_of_interest)
}
\arguments{
\item{gtf_attributes}{the gtf attribute separated by ";"}

\item{att_of_interest}{name of the attributes (e.g. gene_name, gene_biotype)}
}
\value{
attribute of interest in vector format
}
\description{
Extract specific feature of gtf attribute by attribute name.
}
