% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{extract_correlations}
\alias{extract_correlations}
\title{extract_correlations}
\usage{
extract_correlations(mat, num_cov, method = "pearson", filer_genes = F)
}
\arguments{
\item{mat}{matrix of genes x samples}

\item{num_cov}{numeric covariate to correlate against}

\item{method}{used for correlation ('pearson' or 'spearman')}

\item{filer_genes}{Should lowly expressed genes (rowMeans(mat) < 1) be filtered? (default: F)}
}
\value{
table with correlations
}
\description{
Function that streamlines differntial expression of pathways using limma
}
