% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_gff.R
\docType{methods}
\name{extract_GRanges}
\alias{extract_GRanges}
\title{extract_GRanges}
\usage{
extract_GRanges(gff_file)
}
\arguments{
\item{gff_file}{Location of the .gff file}
}
\value{
GRanges object
}
\description{
Generates a GenomicRanges object from a .gff file
}
