% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_gff.R
\docType{methods}
\name{extract_GRanges_from_gtf}
\alias{extract_GRanges_from_gtf}
\title{extract_GRanges_from_gtf}
\usage{
extract_GRanges_from_gtf(gtf_file)
}
\arguments{
\item{gtf_file}{Location of the .gtf file}
}
\value{
GRanges object
}
\description{
Generates a GenomicRanges object from a .gtf file
}
