% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_plots.R
\name{FC_boxplot}
\alias{FC_boxplot}
\title{FC_boxplot}
\usage{
FC_boxplot(
  se,
  mat,
  marker,
  visit_cov = "screening",
  screening_level = "SCREENING",
  cohort_cov = NA,
  subjectid = "USUBJID",
  ylab = marker,
  title = "Fold-Change boxplots",
  ylim = NULL
)
}
\arguments{
\item{se}{Standard summarized experiment generated by MrClean}

\item{mat}{Feature matrix that has the same order as the SE (either gene expression or pathway scores)}

\item{marker}{Feature that is one row in the feature matrix}

\item{visit_cov}{Covariate in the SE that specifies the VISIT (default: screening)}

\item{screening_level}{Level in the visit_cov that specifies screening. All non-screening samples are labeled on-treatment (default: SCREENING)}

\item{cohort_cov}{If there are multiple cohorts there will a pair of boxes will be drawn for each. cohort_cov specifies the covariate in the SE that contains the cohort information (default: NA) in which case only one pair of boxes will be drawn.}

\item{subjectid}{Covariate in the SE that contains the subject ID (default: USUBJID)}

\item{ylab}{Y label of the boxplot - by default this is equal to marker}

\item{title}{Title of the boxplot}
}
\description{
Fold-change boxplot that connects the same subject and shows trends this way
}
