% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_heatmap.R
\name{hcopt}
\alias{hcopt}
\title{hcopt}
\usage{
hcopt(d, HC = NULL, method = "ward.D", members = NULL)
}
\arguments{
\item{d}{a dissimilarity structure as produced by dist.}

\item{HC}{a hierarchical clustering object (default: NULL)}

\item{method}{the agglomeration method to be used.}

\item{members}{NULL or a vector with length size of d. See the 'Details' section fo the hclust function.}
}
\description{
Ward clustering with the optimal leaf ordering algorithm
}
