% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hm.R
\name{plot_hm}
\alias{plot_hm}
\title{plot_hm}
\usage{
plot_hm(
  mat,
  genes,
  main = "",
  col_annot = NA,
  annot_colors = NA,
  value_cap = 3,
  string_cap = 40,
  col = brewer.pal(11, "RdBu")[11:1],
  font.size = 8,
  cluster_cols = T,
  cluster_rows = T,
  show_colnames = T,
  show_rownames = T
)
}
\arguments{
\item{mat}{gene expression matrix}

\item{genes}{genes to include}

\item{main}{title of the heatmap}

\item{col_annot}{data frame that specifies the annotations shown on top of the heatmap. Each row defines the features for a specific column The rows in the data and in the annotation are matched using corresponding column names. Note that color schemes takes into account if variable is continuous or discrete.}

\item{annot_colors}{list for specifying annotation_row and annotation_col track colors manually. It is possible to define the colors for only some of the features.}

\item{value_cap}{(Default: 3) Each expression value will be capped at +/- of this value after row scaling. This corresponds to a cap of 3 standard devivations.}

\item{string_cap}{(default: 40) Row name labels are capped at 40, which is particularly useful for pathways}

\item{col}{(Default: Blue-Red) Color scheme}

\item{font.size}{(Default: 8) Size of the label fonts}

\item{cluster_cols}{(Default: T) boolean value determining whether columns should be clustered or hclust object.}

\item{cluster_rows}{(Default: T) boolean value determining whether rows should be clustered or hclust object.}

\item{show_colnames}{boolean specifying if column names are be shown.}

\item{show_rownames}{boolean specifying if column names are be shown.}
}
\value{
row scaled gene expression matrix
}
\description{
Simple heatmap function that is used to output differential expression results
}
