% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_plots.R
\name{plot_KM}
\alias{plot_KM}
\title{plot_KM}
\usage{
plot_KM(
  stime,
  events,
  marker,
  palette = NULL,
  time_measure = "Days",
  time_label = "Progression free survival",
  main = NULL,
  submain = NULL,
  conf.int = F,
  cex = 1
)
}
\arguments{
\item{stime}{Numeric vector indicating the survival time}

\item{events}{0/1 vector, where 1 corresponds to death or progression}

\item{marker}{groups (factor)}

\item{palette}{color}

\item{time_measure}{(default: Days) y label that is used in the KM plot}

\item{time_label}{(default: PFS) type of survival}

\item{main}{(default: NULL) Title of the plot}

\item{submain}{subtitle of the plot}

\item{conf.int}{confidence interval (default: F)}
}
\description{
Plots a KM curve based on a continuous marker
}
