% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_plots.R
\name{plot_KM_optimal}
\alias{plot_KM_optimal}
\title{plot_KM_optimal}
\usage{
plot_KM_optimal(
  stime,
  events,
  marker,
  palette = NULL,
  time_measure = "Days",
  time_label = "PFS",
  main = NULL,
  conf.int = F,
  HiLoClasses = c("Low", "High"),
  cex = 1
)
}
\arguments{
\item{stime}{Numeric vector indicating the survival time}

\item{events}{0/1 vector, where 1 corresponds to death or progression}

\item{marker}{continuous marker}

\item{palette}{color palette}

\item{time_measure}{(default: Days) y label that is used in the KM plot}

\item{time_label}{(default: PFS) type of survival}

\item{main}{(default: NULL) Title of the plot}

\item{conf.int}{confidence interval (default: F)}

\item{HiLoClasses}{(default: c('Low', 'High')) High and low class names}
}
\description{
Finds an optimal cutpoint for a continuous marker
}
