% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano}
\usage{
plot_volcano(
  diff_genes,
  p.val.cutoff = 0.001,
  p.val.name = "pvalue",
  p.val.lower.limit = 1e-08,
  logfc.name = "log2FoldChange",
  gene.name = "rownames",
  n.top = 20,
  n.max = 50,
  ylab = "-log10(p.val)",
  title = ""
)
}
\arguments{
\item{diff_genes}{data.frame after differential gene analysis: with logFC, gene.name and p.val columns at least}

\item{p.val.cutoff}{cutoff of p.val to show the gene labels.}

\item{p.val.name}{column name for the p value in diff_genes}

\item{p.val.lower.limit}{the lowest p.val limit, if p.val is smaller than it, it will be set to this value to avoid volcano becoming too high.}

\item{logfc.name}{column name for the log fold change in diff_genes}

\item{gene.name}{by default, it uses rownames as gene name, but if there is particular column for gene.name, specifity it here.}

\item{n.top}{(default: 20) number of top differential genes to show the gene labels.}

\item{n.max}{(default: 50) maximum number of genes to show on the volcano to avoid over-populating the volcano.
A gene will be labeled if it either meets the p.val.cutoff or meets the n.top criteria.}

\item{ylab}{label of y-axis}

\item{title}{title of the plot}
}
\value{
a ggplot object.
}
\description{
Volcano plot to show top differential expressed genes
}
\examples{
\dontrun{
plot_volcano(diff_genes, p.val.cutoff = 1e-4, logfc.name = "log2FoldChange", p.val.name = "pvalue")
}
}
