% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_function.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA}
\usage{
plotPCA(
  cov,
  pca,
  var_explained,
  covs,
  pos = "bottomleft",
  col_set_factor = "Set1",
  col_set_numeric = "Blues"
)
}
\arguments{
\item{cov}{Covariate that should be plotted}

\item{pca}{Pre-computed pca using the prcomp functions}

\item{var_explained}{Variance explained by the single principle components}

\item{covs}{Data frame (or SummarizedExperiment) containing the covariate that should be plotted}

\item{pos}{Position of the legend}

\item{col_set_factor}{RColorBrewer color scheme for factors}

\item{col_set_numeric}{RColorBrewer color scheme for numbers}
}
\description{
Plots a PCA and colors the dots according to the classes of a covariate
}
