% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_plots.R
\name{project_pathway_space}
\alias{project_pathway_space}
\title{project_pathway_space}
\usage{
project_pathway_space(mat, gene_sets, method = "means")
}
\arguments{
\item{mat}{gene expression matrix with ncol samples and nrow genes}

\item{gene_sets}{list of gene set vectors}

\item{method}{(default: means) currently only means and 'geometric_means' is implemented}
}
\value{
Projected pathway x samples matrix
}
\description{
Takes a gene expression matrix and projects it into pathway space
}
