% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_plots.R
\name{rank_by_wilcoxon}
\alias{rank_by_wilcoxon}
\title{rank_by_wilcoxon}
\usage{
rank_by_wilcoxon(mat, group)
}
\arguments{
\item{mat}{matrix of features x samples}

\item{group}{identify the sample groups that should be compared}
}
\description{
takes a matrix of features by samples, splits samples by a group designation and does a wilcoxon test for each measurement. Returns a table of features ranked by p-value.
}
