% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liver_confounding.R
\docType{methods}
\name{removeLiverConfounder}
\alias{removeLiverConfounder}
\title{Liver confounder removal}
\usage{
removeLiverConfounder(
  dat,
  plot = T,
  method = c("mean", "gsva"),
  se_matrix = "final"
)
}
\arguments{
\item{dat}{Gene expression matrix - rows are genes and columns are samples.}

\item{plot}{Flag indicating whether a PCA should be plotted.}

\item{method}{Method used to estimate the liver score ('mean' or 'gsva')}

\item{se_matrix}{indicates which matrix to use in case a SE is passed. By default this is final, but can usually also be log2_tmm_cpm}
}
\value{
Vector of gene scores indicating the levels of liver confounding
}
\description{
Removes a liver confounder in a gene expression dataset
}
