% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hm.R
\name{rowScale}
\alias{rowScale}
\title{rowScale}
\usage{
rowScale(x)
}
\arguments{
\item{x}{gene expression matrix}
}
\value{
row scaled gene expression matrix
}
\description{
Simple function to perform row scaling
}
