% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{run_diff_pathways}
\alias{run_diff_pathways}
\title{run_diff_pathways}
\usage{
run_diff_pathways(
  pathways,
  se,
  group = NULL,
  main = NULL,
  design_mat = NULL,
  coef = NULL,
  plot_hm = T,
  col_annot = NA,
  annot_colors = NA,
  num_hm_genes = 30,
  show_colnames = T,
  show_rownames = T,
  cluster_cols = T,
  cluster_rows = T
)
}
\arguments{
\item{pathways}{matrix of gene sets x samples}

\item{se}{SummarizedExperiment}

\item{group}{group to do the differential expression with}

\item{main}{title of the heatmap}

\item{design_mat}{custom design matrix if a more complicated model is needed}

\item{coef}{coeffient to use for the likelihood ratio test when specifying a custom design matrix}

\item{plot_hm}{Should a heatmap be plotted? (default: T)}

\item{col_annot}{Heatmap annotation. Data frame that specifies the annotations shown on top of the heatmap. Each row defines the features for a specific column The rows in the data and in the annotation are matched using corresponding column names. Note that color schemes takes into account if variable is continuous or discrete.}

\item{annot_colors}{Heatmap annotation colors. List for specifying annotation_row and annotation_col track colors manually. It is possible to define the colors for only some of the features.}

\item{num_hm_genes}{Number of genes to be used in the heatmap}

\item{show_colnames}{boolean specifying if column names are be shown.}

\item{show_rownames}{boolean specifying if column names are be shown.}

\item{cluster_cols}{(Default: T) boolean value determining whether columns should be clustered or hclust object.}

\item{cluster_rows}{(Default: T) boolean value determining whether rows should be clustered or hclust object.}
}
\value{
data frame of differentially expressed pathways
}
\description{
Function that streamlines differntial expression of pathways using limma
}
