% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{run_limma}
\alias{run_limma}
\title{run_limma}
\usage{
run_limma(
  se,
  group = NULL,
  main = "",
  col_annot = NA,
  annot_colors = NA,
  num_hm_genes = 30,
  design_mat = NULL,
  assay_name_counts = "log2_tmm_cpm",
  assay_name_final = "final",
  coef = NULL,
  cluster_cols = T,
  cluster_rows = T,
  show_colnames = T,
  show_rownames = T
)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{group}{group to do the differential expression with}

\item{main}{title of the heatmap}

\item{col_annot}{Heatmap annotation. Data frame that specifies the annotations shown on top of the heatmap. Each row defines the features for a specific column The rows in the data and in the annotation are matched using corresponding column names. Note that color schemes takes into account if variable is continuous or discrete.}

\item{annot_colors}{Heatmap annotation colors. List for specifying annotation_row and annotation_col track colors manually. It is possible to define the colors for only some of the features.}

\item{num_hm_genes}{Number of genes to be used in the heatmap}

\item{design_mat}{custom design matrix if a more complicated model is needed}

\item{assay_name_counts}{(Default: "raw_counts") the assay used for gene expression counts (i.e. \code{assays(se)[[assay_name_counts]]})}

\item{assay_name_final}{(Default: "final") the assay used for final plotting (i.e. \code{assays(se)[[assay_name_final]]})}

\item{coef}{coefficient to use for the likelihood ratio test when specifying a custom design matrix}

\item{cluster_cols}{(Default: T) boolean value determining whether columns should be clustered or hclust object.}

\item{cluster_rows}{(Default: T) boolean value determining whether rows should be clustered or hclust object.}

\item{show_colnames}{boolean specifying if column names are be shown.}

\item{show_rownames}{boolean specifying if column names are be shown.}
}
\value{
table of differentially expressed genes
}
\description{
Function that streamlines differntial expression using limma
}
