% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_plots.R
\name{scatter}
\alias{scatter}
\title{scatter}
\usage{
scatter(
  marker1,
  marker2,
  col = "#444444",
  annot_colors = NULL,
  trendline = T,
  trendline_col = "#3182bd",
  add_correlation = T,
  correlation = "spearman",
  cex.pt = 1.5,
  pch = 16,
  xlab = "x",
  ylab = "y",
  main = NULL,
  plot_legend = T,
  legend_cex = 0.8,
  legend_pos = "topright",
  cex = 1,
  ylim = NULL,
  xlim = NULL
)
}
\arguments{
\item{marker1}{Biomarker to show on the x axis}

\item{marker2}{Biomarker to show on the y axis}

\item{col}{color of the scatter, can be a single color or a factorial (default: '#444444')}

\item{annot_colors}{Named list of colors corresponding to the levels of the color factorial (default: NULL)}

\item{trendline}{Draw a trendline on the plot (default: T)}

\item{trendline_col}{Color of the trendline (default: '#3182bd')}

\item{add_correlation}{Add a correlation value and p value on top of the plot (default: T)}

\item{correlation}{Type of correlation to be used (default: 'spearman')}

\item{cex.pt}{pch(default:  16)}

\item{xlab}{label of the X axis (default: 'x')}

\item{ylab}{label of the y axis (default: 'y')}

\item{main}{Header (default: NULL)}

\item{plot_legend}{Should a legend be drawn if there are multiple colors used? (Default: T)}

\item{legend_cex}{Text size in the legend (default: 0.8)}

\item{legend_pos}{Position of the legend (default: 'topright')}

\item{cex}{font_sizes}
}
\description{
Standard scatter plot with correlation test, legend and trendline
}
