% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_plots.R
\name{standard_boxp}
\alias{standard_boxp}
\title{standard_boxp}
\usage{
standard_boxp(
  sample_list,
  col = "#999999",
  ylab = "y",
  main = NULL,
  wilcox = TRUE,
  jitter_method = "jitter",
  overlay_dot_col = "#444444",
  las = 1,
  cex = 1,
  cex.axis = 1,
  cex.dot = 1,
  ylim = NULL
)
}
\arguments{
\item{sample_list}{list of biomarker vectors to be plotted}

\item{col}{(Default: '#999999') Color of boxplot. Can also be a color vector}

\item{ylab}{(Default: 'y') y axis label}

\item{main}{Title of the plot (default: NULL)}

\item{wilcox}{(Default: TRUE) Should a Wilcoxon p-value be calculated?}

\item{jitter_method}{(Default: jitter) Stripchart overlay of the boxplots. Can also be overplot and stack}

\item{overlay_dot_col}{(Default: "#444444") Stripchart color}

\item{las}{(Default: 1) Label orientation}

\item{cex}{font_sizes}
}
\description{
Standard boxplot with Wilcoxon test
}
