% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_plots.R
\name{tabify}
\alias{tabify}
\title{tabify}
\usage{
tabify(factor1, factor2, by_row = F)
}
\arguments{
\item{factor1}{Factorial}

\item{factor2}{Factorial}

\item{by_row}{(default: F) Should the percentage calculated by row or by column}
}
\description{
Function that takes two factorials, generates a table breakdown and calculates percentages
}
