% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmm_normalize.R
\docType{methods}
\name{tmm_normalize}
\alias{tmm_normalize}
\title{tmm_normalize}
\usage{
tmm_normalize(se)
}
\arguments{
\item{se}{SummarizedExperiment}
}
\value{
SummarizedExperiment that contains the normalized data
}
\description{
Standard function to normalize a RNAseq cube
}
